`set(xpath,val)` - extract and raise libyang errors
By overriding the libyang error callback we can raise error messages if
data `set(xpath, value)` is subsequently called violating the
constraints.

If data does not yet exist in the data tree then the value *MUST* be valid.
However, if does already exist in the data tree the libyang-cffi call to
set will not throw any exceptions as the test of success is based on the
presence of the `lyd_node` not the update of a `lyd_node`.