Fetch non-existing presence containers on demand.
Using 'tabs' at the top of the page is a littel bit complex:

If we have a yang model

```
container a{
  t:yangui-tab "First Tab";
}
container b{
  t:yangui-tab "Second Tab";
}
container c{
}
leaf id {
}
```

We would have two explicit tabs, and probably an uncategorised tab

If we have a yang model

```
list mylist {
 key id;

  container a{
    t:yangui-tab "First Tab";
  }
  container b{
    t:yangui-tab "Second Tab";
  }
  container c{
  }
  leaf id {
  }
}
```

Now it's unclear what we would do - it's not as simple as putting things
in tabs based on annotations as the relationship between list-element
and containers is easily broken.

It may be possible to do something like a 'dynamic tab' created based on
user interaction.

It may be that the optimisation of only adding <div>'s for presence
containers with contents may be enough of an optimisation (although if a
data-tree has data spread across a large model it may be marginal)